# [stdlib] Simplify definition of Pos.add and Pos.add_carry