--- a/doc/changelog/10-standard-library/13936-simpler-pos-add.rst
+++ b/doc/changelog/10-standard-library/13936-simpler-pos-add.rst
@@ -1,0 +1,4 @@
+- **Changed:**
+  Simplified definition of `Pos.add` and `Pos.add_carry` to not use mutual recursion
+  (`#13936 <https://github.com/coq/coq/pull/13936>`_,
+  by Li-yao Xia).