@ppedrot 2021-03-12
There might be an efficiency reason for that choice. The new version traverses the number twice, while the old one did it only once.