@Lysxia 2021-03-12
That's true. I suppose some people care about such computations being fast in proofs.

Just for fun, on a very artificial benchmark I measured a 20% slowdown.

Source below.

```
vm_compute
Before: 0.355 secs
After: 0.42 secs

cbv
Before: 5.194 secs
After: 6.566 secs
```

```coq
From Coq Require Import PArith.

Fixpoint add (x y : positive) {struct x} : positive :=
  match x with
  | (p~1)%positive =>
      match y with
      | (q~1)%positive => ((Pos.succ (add p q))~0)%positive
      | (q~0)%positive => ((add p q)~1)%positive
      | 1%positive => ((Pos.succ p)~0)%positive
      end
  | (p~0)%positive =>
      match y with
      | (q~1)%positive => ((add p q)~1)%positive
      | (q~0)%positive => ((add p q)~0)%positive
      | 1%positive => (p~1)%positive
      end
  | 1%positive =>
      match y with
      | (q~1)%positive => ((Pos.succ q)~0)%positive
      | (q~0)%positive => (q~1)%positive
      | 1%positive => 2%positive
      end
  end.

Notation a := Pos.add.

Fixpoint vv (n : positive) (q0 : positive) : positive :=
  match n with
  | xH => q0
  | xO n => a (vv n q0) (vv n (a q0 n))
  | xI n => Pos.succ (a (vv n q0) (vv n (a q0 n)))
  end.

Notation b := add.

Fixpoint ww (n : positive) (q0 : positive) : positive :=
  match n with
  | xH => q0
  | xO n => b (ww n q0) (ww n (b q0 n))
  | xI n => Pos.succ (b (ww n q0) (ww n (b q0 n)))
  end.

Time Eval vm_compute in (vv 1000000 1000).
Time Eval vm_compute in (ww 1000000 1000).

Time Eval cbv in (vv 1000000 1000).
Time Eval cbv in (ww 1000000 1000).
```